Initial Version
Contains the jupyter notebook with functions to read sensor output files, denoise and process them, compute new parameters from them (velocity and angle) and visualize them for comparison/future analysis. The sample data in exp1, is for 1 individual walking with 5 sensors on (head sensor was placed on the waist and not acctually on the head).